converter-saas/backend/requirements.txt
# Core Dependencies
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0

# File Processing
Pillow>=10.0,<11.0
python-magic-bin>=0.4.14;sys_platform=='win32'
python-magic>=0.4.27;sys_platform!='win32'
ffmpeg-python>=0.2.0,<0.3.0

# PDF Processing
PyPDF2>=3.0,<4.0
pdf2image>=1.16,<2.0
python-docx>=1.0,<2.0
openpyxl>=3.1,<4.0
python-pptx>=0.6.21,<0.7.0
markdown>=3.5,<4.0
WeasyPrint>=60.0,<61.0
reportlab>=4.0,<5.0

# Image Processing (additional)
cairosvg>=2.7,<3.0
pillow-heif>=0.13,<0.14

# Async Task Queue (future)
celery>=5.3,<6.0
redis>=5.0,<6.0

# Utilities
python-dotenv>=1.0,<2.0
uuid>=1.30
